backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
bcrypt
python-jose[cryptography]
pydantic[email]
